descrypt-ztex: Exclude colliding test vectors
Our self-test isn't prepared to handle this for formats that remap indices.
It finds just one index for the hash with a loop doing cmp_one(), but with
a collision this may end up corresponding to the other plaintext, so the
check of get_key() return value against the expected plaintext fails.

In actual cracking, descrypt-ztex --keep-guessing is able to find all 6
plaintexts just fine, with a 6-line wordlist containing them in full (so
running without mask) or truncated (plus on-device mask).

In contrast, descrypt-opencl only finds 3 plaintexts that come first in the
wordlist, but seems to have no problem passing self-test with collisions
included, perhaps due to its clear_keys().