HMAC: Optimize opad XOR (in C code only, not OpenCL)
There are several possible ways to construct HMAC's ipad and opad:
1. Initialize them (or one buffer to be reused as both) to key with NUL
   padding, then XOR the constants over that.
2. Initialize them to the constants, then XOR the key over the initial
   portions (until key length).
3. Compute the initial portions right away, then pass the rest of the
   constant data into separate *_Update() calls.
4. Combinations of the above and other variations (e.g., partial memset(),
   doing it out of a loop, etc.)

This commit optimizes existing implementations of 1 and 3 by XOR'ing the
one reused pad buffer with XOR of the two constants instead of re-reading
the key and XOR'ing that with the second constant.  Some implementations
already had this optimization, and so are not touched by this commit.

Explicit usage of larger-than-byte XORs may be even more important, but
isn't implemented with this minimal-change commit (would be non-trivial
with arbitrary sizes and alignment of keys).  Some implementations already
have this optimization.

This commit does not touch existing implementations of 2 and 4.  Those may
(or may not) actually be as good as or even more optimal than optimized 1,
such as due to efficiency of memset() and typically low key lengths.

This commit reduces john .text size by ~400 bytes in a Linux/x86_64 build.