pcap2john: Add SNMPv3 parser
This adds support for extracting SNMPv3 USM hashes. The parser handles
both authentication-only (authNoPriv) and authentication with privacy
(authPriv) modes.

Hashes are printed in $SNMPv3$ format, with authProto set to 0 to allow
John to try both MD5 and SHA1. If the authProto is known the 0
can be changed to 1 (MD5) or 2 (SHA1). This is not done automatically.

Signed-off-by: Albert Veli <albert.veli@gmail.com>