SL3 formats: Refuse to accept a 15 digit IMEI with incorrect Luhn digit
That digit isn't used by the algo (and only specifying 14 digits without
Luhn is and was accepted) but this change will somewhat lower the risk
of misidentifying other hashes as SL3.

There were actually a mockup test vector featuring an incorrect Luhn so
I had to correct that as well.

See #5767