OpenCL: Self-test with default LWS at device maximum
Also with a GWS of 2x that LWS.  The new figures are better at triggering
bugs. If a kernel needs a lower LWS than device max. our code already
handles that.

We previously had it as LWS=7 GWS=49 for speed (and for checking
heuristics that could bug out on non-log2 values) but that was
introduced before our autotune was sped up with orders of magnitude
and those heuristics has been stable for many years.

Like before, the self-test will obey any given lws/gws options or
environment variables (in that case even when they are past limits).

Closes #5822