Fix bugs in a few formats' plugin logic
Some formats had an #elif or #else creatively skipped in the plugin logic,
messing things up royally and slowing down the build process. It's not a
good idea to write that stuff in any other way, or put anything before it
(with very few exceptions, one being an outermost #if HAVE_OPENCL).